refactor: integrate persistent controller with enhanced watchlist
